2263: remove TB's GL abstraction layer and just use glew
