+++ cmake/TrenchBroomTest.cmake
@@ -13,7 +13,6 @@ FILE(GLOB_RECURSE BENCHMARK_SOURCE
 # Re-use some of TrenchBroom-Test (e.g. main()) in TrenchBroom-Benchmark
 LIST(APPEND BENCHMARK_SOURCE 
 	"${TEST_SOURCE_DIR}/RunAllTests.cpp"
-	"${TEST_SOURCE_DIR}/GLInit.cpp"
 	"${TEST_SOURCE_DIR}/View/GetVersion.cpp")
 
 ADD_EXECUTABLE(TrenchBroom-Test ${TEST_SOURCE} $<TARGET_OBJECTS:common>)
