+++ common/src/View/GLContext.h
@@ -22,6 +22,7 @@
 
 #include "SharedPointer.h"
 
+#include <GL/glew.h>
 #include <wx/glcanvas.h>
 
 namespace TrenchBroom {
